Refactor backed up resources
