+++ service/discovery/azure/azure.go
@@ -52,7 +52,6 @@ import (
 	"github.com/sirupsen/logrus"
 
 	"clouditor.io/clouditor/api/discovery"
-	"clouditor.io/clouditor/internal/constants"
 	"clouditor.io/clouditor/internal/util"
 	"clouditor.io/clouditor/voc"
 )
@@ -89,12 +88,6 @@ var (
 // TODO (anatheka): Merge discoverers:
 // - fix backupStorage
 // - fix tests
-
-// TODO (anatheka): Discoveres:
-// NewAzureResourceGroupDiscovery
-// NewAzureComputeDiscovery
-// NewAzureStorageDiscovery
-// NewAzureNetworkDiscovery
 
 func (*azureDiscovery) Name() string {
 	return "Azure"
@@ -150,8 +143,15 @@ type azureDiscovery struct {
 	defenderProperties  map[string]*defenderProperties
 }
 
+type defenderProperties struct {
+	monitoringLogDataEnabled bool
+	securityAlertsEnabled    bool
+}
+
 type backup struct {
-	backup         map[string][]*voc.Backup
+	// backup is a list of all voc.Backup objects
+	backup map[string][]*voc.Backup
+	// backupStorages is a list of all backed up voc.Storage (ObjectStorage, BlockStorage) objects
 	backupStorages []voc.IsCloudResource
 }
 
@@ -209,10 +209,10 @@ func NewAzureDiscovery(opts ...DiscoveryOption) discovery.Discoverer {
 }
 
 // List discovers the following Azure resources types:
+// - ResourceGroup resource
 // - Storage resource
 // - Compute resource
 // - Network resource
-// - ResourceGroup resource
 func (d *azureDiscovery) List() (list []voc.IsCloudResource, err error) {
 	if err = d.authorize(); err != nil {
 		return nil, fmt.Errorf("%s: %w", ErrCouldNotAuthenticate, err)
@@ -383,11 +383,6 @@ func NewAuthorizer() (*azidentity.DefaultAzureCredential, error) {
 	return cred, nil
 }
 
-type defenderProperties struct {
-	monitoringLogDataEnabled bool
-	securityAlertsEnabled    bool
-}
-
 // discoverDefender discovers Defender for X services and returns a map with the following properties for each defender type
 // * monitoringLogDataEnabled
 // * securityAlertsEnabled
@@ -398,10 +393,6 @@ func (d *azureDiscovery) discoverDefender() (map[string]*defenderProperties, err
 	// initialize defender client
 	if err := d.initDefenderClient(); err != nil {
 		return nil, fmt.Errorf("could not initialize defender client: %w", err)
-	}
-
-	if d.clients.defenderClient == nil {
-		return nil, errors.New("defenderClient not set")
 	}
 
 	// List all pricings to get the enabled Defender for X
@@ -427,170 +418,23 @@ func (d *azureDiscovery) discoverDefender() (map[string]*defenderProperties, err
 	return pricings, nil
 }
 
-// discoverBackupVaults receives all backup vaults in the subscription.
-// Since the backups for storage and compute are discovered together, the discovery is performed here and results are stored in the azureDiscovery receiver.
-func (d *azureDiscovery) discoverBackupVaults() error {
-
-	if d.backupMap != nil && len(d.backupMap) > 0 {
-		log.Debug("Backup Vaults already discovered.")
-		return nil
-	}
-
-	// initialize backup vaults client
-	if err := d.initBackupVaultsClient(); err != nil {
-		return err
-	}
-
-	// initialize backup instances client
-	if err := d.initBackupInstancesClient(); err != nil {
-		return err
-	}
-
-	// initialize backup policies client
-	if err := d.initBackupPoliciesClient(); err != nil {
-		return err
-	}
-
-	if d.clients.backupVaultClient == nil || d.clients.backupInstancesClient == nil {
-		return errors.New("backupVaultClient and/or backupInstancesClient missing")
-	}
-
-	// List all backup vaults
-	err := listPager(d,
-		d.clients.backupVaultClient.NewGetInSubscriptionPager,
-		d.clients.backupVaultClient.NewGetInResourceGroupPager,
-		func(res armdataprotection.BackupVaultsClientGetInSubscriptionResponse) []*armdataprotection.BackupVaultResource {
-			return res.Value
-		},
-		func(res armdataprotection.BackupVaultsClientGetInResourceGroupResponse) []*armdataprotection.BackupVaultResource {
-			return res.Value
-		},
-		func(vault *armdataprotection.BackupVaultResource) error {
-			instances, err := d.discoverBackupInstances(resourceGroupName(util.Deref(vault.ID)), util.Deref(vault.Name))
-			if err != nil {
-				err := fmt.Errorf("could not discover backup instances: %v", err)
-				return err
-			}
-
-			for _, instance := range instances {
-				dataSourceType := util.Deref(instance.Properties.DataSourceInfo.DatasourceType)
-
-				// Get retention from backup policy
-				policy, err := d.clients.backupPoliciesClient.Get(context.Background(), resourceGroupName(*vault.ID), *vault.Name, backupPolicyName(*instance.Properties.PolicyInfo.PolicyID), &armdataprotection.BackupPoliciesClientGetOptions{})
-				if err != nil {
-					err := fmt.Errorf("could not get backup policy '%s': %w", *instance.Properties.PolicyInfo.PolicyID, err)
-					log.Error(err)
-					continue
-				}
-
-				// TODO(all):Maybe we should differentiate the backup retention period for different resources, e.g., disk vs blobs (Metrics)
-				retention := policy.BaseBackupPolicyResource.Properties.(*armdataprotection.BackupPolicy).PolicyRules[0].(*armdataprotection.AzureRetentionRule).Lifecycles[0].DeleteAfter.(*armdataprotection.AbsoluteDeleteOption).GetDeleteOption().Duration
-
-				resp, err := d.handleInstances(vault, instance)
-				if err != nil {
-					err := fmt.Errorf("could not handle instance")
-					return err
-				}
-
-				// Check if map entry already exists
-				_, ok := d.backupMap[dataSourceType]
-				if !ok {
-					d.backupMap[dataSourceType] = &backup{
-						backup: make(map[string][]*voc.Backup),
-					}
-				}
-
-				// Store voc.Backup in backupMap
-				d.backupMap[dataSourceType].backup[util.Deref(instance.Properties.DataSourceInfo.ResourceID)] = []*voc.Backup{
-					{
-						Enabled:         true,
-						RetentionPeriod: retentionDuration(util.Deref(retention)),
-						Storage:         voc.ResourceID(util.Deref(instance.ID)),
-						TransportEncryption: &voc.TransportEncryption{
-							Enabled:    true,
-							Enforced:   true,
-							Algorithm:  constants.TLS,
-							TlsVersion: constants.TLS1_2, // https://learn.microsoft.com/en-us/azure/backup/transport-layer-security#why-enable-tls-12 (Last access: 04/27/2023)
-						},
-					},
-				}
-
-				d.backupMap[dataSourceType].backupStorages = append(d.backupMap[dataSourceType].backupStorages, resp)
-			}
-			return nil
-		})
-
-	if err != nil {
-		return err
-	}
-
-	return nil
+// resourceGroupName returns the resource group name of a given Azure ID
+func resourceGroupName(id string) string {
+	return strings.Split(id, "/")[4]
 }
 
-// backupsEmptyCheck checks if the backups list is empty and returns voc.Backup with enabled = false.
-func backupsEmptyCheck(backups []*voc.Backup) []*voc.Backup {
-	if len(backups) == 0 {
-		return []*voc.Backup{
-			{
-				Enabled:         false,
-				RetentionPeriod: -1,
-				Interval:        -1,
-			},
-		}
+func resourceGroupID(ID *string) voc.ResourceID {
+	// split according to "/"
+	s := strings.Split(util.Deref(ID), "/")
+
+	// We cannot really return an error here, so we just return an empty string
+	if len(s) < 5 {
+		return ""
 	}
 
-	return backups
-}
+	id := strings.Join(s[:5], "/")
 
-func (d *azureDiscovery) handleInstances(vault *armdataprotection.BackupVaultResource, instance *armdataprotection.BackupInstanceResource) (resource voc.IsCloudResource, err error) {
-	if vault == nil || instance == nil {
-		return nil, ErrVaultInstanceIsEmpty
-	}
-
-	raw, err := voc.ToStringInterface([]interface{}{instance, vault})
-	if err != nil {
-		log.Error(err)
-	}
-
-	if *instance.Properties.DataSourceInfo.DatasourceType == "Microsoft.Storage/storageAccounts/blobServices" {
-		resource = &voc.ObjectStorage{
-			Storage: &voc.Storage{
-				Resource: &voc.Resource{
-					ID:           voc.ResourceID(*instance.ID),
-					Name:         *instance.Name,
-					CreationTime: 0,
-					GeoLocation: voc.GeoLocation{
-						Region: *vault.Location,
-					},
-					Labels:    nil,
-					ServiceID: d.csID,
-					Type:      voc.ObjectStorageType,
-					Parent:    resourceGroupID(instance.ID),
-					Raw:       raw,
-				},
-			},
-		}
-	} else if *instance.Properties.DataSourceInfo.DatasourceType == "Microsoft.Compute/disks" {
-		resource = &voc.BlockStorage{
-			Storage: &voc.Storage{
-				Resource: &voc.Resource{
-					ID:           voc.ResourceID(*instance.ID),
-					Name:         *instance.Name,
-					ServiceID:    d.csID,
-					CreationTime: 0,
-					Type:         voc.BlockStorageType,
-					GeoLocation: voc.GeoLocation{
-						Region: *vault.Location,
-					},
-					Labels: nil,
-					Parent: resourceGroupID(instance.ID),
-					Raw:    raw,
-				},
-			},
-		}
-	}
-
-	return
+	return voc.ResourceID(id)
 }
 
 // retentionDuration returns the retention string as time.Duration
@@ -613,50 +457,6 @@ func retentionDuration(retention string) time.Duration {
 	duration := time.Duration(time.Duration(d) * time.Hour * 24)
 
 	return duration
-}
-
-// discoverBackupInstances retrieves the instances in a given backup vault.
-// Note: It is only possible to backup a storage account with all containers in it.
-func (d *azureDiscovery) discoverBackupInstances(resourceGroup, vaultName string) ([]*armdataprotection.BackupInstanceResource, error) {
-	var (
-		list armdataprotection.BackupInstancesClientListResponse
-		err  error
-	)
-
-	if resourceGroup == "" || vaultName == "" {
-		return nil, errors.New("missing resource group and/or vault name")
-	}
-
-	// List all instances in the given backup vault
-	listPager := d.clients.backupInstancesClient.NewListPager(resourceGroup, vaultName, &armdataprotection.BackupInstancesClientListOptions{})
-	for listPager.More() {
-		list, err = listPager.NextPage(context.TODO())
-		if err != nil {
-			err = fmt.Errorf("%s: %v", ErrGettingNextPage, err)
-			return nil, err
-		}
-	}
-
-	return list.Value, nil
-}
-
-// resourceGroupName returns the resource group name of a given Azure ID
-func resourceGroupName(id string) string {
-	return strings.Split(id, "/")[4]
-}
-
-func resourceGroupID(ID *string) voc.ResourceID {
-	// split according to "/"
-	s := strings.Split(util.Deref(ID), "/")
-
-	// We cannot really return an error here, so we just return an empty string
-	if len(s) < 5 {
-		return ""
-	}
-
-	id := strings.Join(s[:5], "/")
-
-	return voc.ResourceID(id)
 }
 
 // backupPolicyName returns the backup policy name of a given Azure ID
