+++ service/discovery/azure/azure_test.go
@@ -38,14 +38,12 @@ import (
 	"time"
 
 	"clouditor.io/clouditor/api/discovery"
-	"clouditor.io/clouditor/internal/constants"
 	"clouditor.io/clouditor/internal/testdata"
 	"clouditor.io/clouditor/internal/util"
 	"clouditor.io/clouditor/voc"
 	"github.com/Azure/azure-sdk-for-go/sdk/azcore"
 	"github.com/Azure/azure-sdk-for-go/sdk/azcore/arm"
 	"github.com/Azure/azure-sdk-for-go/sdk/azcore/policy"
-	"github.com/Azure/azure-sdk-for-go/sdk/resourcemanager/dataprotection/armdataprotection"
 	"github.com/Azure/azure-sdk-for-go/sdk/resourcemanager/storage/armstorage"
 	"github.com/Azure/azure-sdk-for-go/sdk/resourcemanager/subscription/armsubscription"
 	"github.com/stretchr/testify/assert"
@@ -599,239 +597,6 @@ func NewMockAzureDiscovery(transport policy.Transporter, opts ...DiscoveryOption
 	return d
 }
 
-// TODO(anatheka): Update test
-func Test_azureDiscovery_discoverBackupVaults_Storage(t *testing.T) {
-	type fields struct {
-		azureDiscovery            *azureDiscovery
-		clientBackupVault         bool
-		clientBackupInstance      bool
-		emptyClientBackupInstance bool
-		clientBackupPolicy        bool
-	}
-	tests := []struct {
-		name    string
-		fields  fields
-		want    assert.ValueAssertionFunc
-		wantErr assert.ErrorAssertionFunc
-	}{
-		{
-			name: "all clients missing",
-			fields: fields{
-				azureDiscovery:       NewMockAzureDiscovery(newMockStorageSender()),
-				clientBackupVault:    false,
-				clientBackupInstance: false,
-				clientBackupPolicy:   false,
-			},
-			want: nil,
-			wantErr: func(t assert.TestingT, err error, i ...interface{}) bool {
-				return assert.ErrorContains(t, err, "backupVaultClient and/or backupInstancesClient missing")
-			},
-		},
-		{
-			name: "backup instance client missing",
-			fields: fields{
-				azureDiscovery:       NewMockAzureDiscovery(newMockStorageSender()),
-				clientBackupVault:    true,
-				clientBackupInstance: false,
-				clientBackupPolicy:   false,
-			},
-			want: nil,
-			wantErr: func(t assert.TestingT, err error, i ...interface{}) bool {
-				return assert.ErrorContains(t, err, "backupVaultClient and/or backupInstancesClient missing")
-			},
-		},
-		{
-			name: "backup instance client empty",
-			fields: fields{
-				azureDiscovery:            NewMockAzureDiscovery(newMockStorageSender()),
-				clientBackupVault:         true,
-				emptyClientBackupInstance: true,
-				clientBackupPolicy:        false,
-			},
-			want: nil,
-			wantErr: func(t assert.TestingT, err error, i ...interface{}) bool {
-				return assert.ErrorContains(t, err, "could not discover backup instances:")
-			},
-		},
-
-		{
-			name: "Happy path",
-			fields: fields{
-				azureDiscovery:       NewMockAzureDiscovery(newMockStorageSender()),
-				clientBackupVault:    true,
-				clientBackupInstance: true,
-				clientBackupPolicy:   true,
-			},
-			want: func(tt assert.TestingT, i1 interface{}, i2 ...interface{}) bool {
-				d, ok := i1.(*azureDiscovery)
-				if !assert.True(tt, ok) {
-					return false
-				}
-
-				want := []*voc.Backup{{
-					RetentionPeriod: Duration7Days,
-					Enabled:         true,
-					Storage:         voc.ResourceID("/subscriptions/00000000-0000-0000-0000-000000000000/resourceGroups/res1/providers/Microsoft.DataProtection/backupVaults/backupAccount1/backupInstances/account1-account1-22222222-2222-2222-2222-222222222222"),
-					TransportEncryption: &voc.TransportEncryption{
-						Enforced:   true,
-						Enabled:    true,
-						TlsVersion: constants.TLS1_2,
-						Algorithm:  constants.TLS,
-					},
-				},
-				}
-
-				got := d.backupMap[DataSourceTypeStorageAccountObject].backup["/subscriptions/00000000-0000-0000-0000-000000000000/resourceGroups/res1/providers/Microsoft.Storage/storageAccounts/account1"]
-
-				return assert.Equal(t, want, got)
-
-			},
-			wantErr: assert.NoError,
-		},
-	}
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			// d := &azureDiscovery{
-			// 	azureDiscovery: tt.fields.azureDiscovery,
-			// }
-			d := tt.fields.azureDiscovery
-
-			// Set clients if needed
-			if tt.fields.clientBackupVault {
-				// initialize backup vaults client
-				_ = d.initBackupVaultsClient()
-			}
-			if tt.fields.clientBackupInstance {
-				// initialize backup instances client
-				_ = d.initBackupInstancesClient()
-			}
-			if tt.fields.clientBackupPolicy {
-				// initialize backup policies client
-				_ = d.initBackupPoliciesClient()
-			}
-
-			// Set empty client if needed
-			if tt.fields.emptyClientBackupInstance {
-				// empty backup instances client
-				d.clients.backupInstancesClient = &armdataprotection.BackupInstancesClient{}
-			}
-
-			err := d.discoverBackupVaults()
-
-			tt.wantErr(t, err)
-
-			if tt.want != nil {
-				tt.want(t, d)
-			}
-		})
-	}
-}
-
-// TODO(anatheka): Update test
-func Test_azureDiscovery_discoverBackupVaults_Compute(t *testing.T) {
-	type fields struct {
-		azureDiscovery       *azureDiscovery
-		clientBackupVault    bool
-		clientBackupInstance bool
-		clientBackupPolicy   bool
-	}
-	tests := []struct {
-		name    string
-		fields  fields
-		want    assert.ValueAssertionFunc
-		wantErr assert.ErrorAssertionFunc
-	}{
-		{
-			name: "both clients missing",
-			fields: fields{
-				azureDiscovery:       NewMockAzureDiscovery(newMockComputeSender()),
-				clientBackupVault:    false,
-				clientBackupInstance: false,
-				clientBackupPolicy:   false,
-			},
-			want: nil,
-			wantErr: func(t assert.TestingT, err error, i ...interface{}) bool {
-				return assert.ErrorContains(t, err, "backupVaultClient and/or backupInstancesClient missing")
-			},
-		},
-		{
-			name: "backup instance client missing",
-			fields: fields{
-				azureDiscovery:       NewMockAzureDiscovery(newMockComputeSender()),
-				clientBackupVault:    true,
-				clientBackupInstance: false,
-				clientBackupPolicy:   false,
-			},
-			want: nil,
-			wantErr: func(t assert.TestingT, err error, i ...interface{}) bool {
-				return assert.ErrorContains(t, err, "backupVaultClient and/or backupInstancesClient missing")
-			},
-		},
-		{
-			name: "Happy path",
-			fields: fields{
-				azureDiscovery:       NewMockAzureDiscovery(newMockComputeSender()),
-				clientBackupVault:    true,
-				clientBackupInstance: true,
-				clientBackupPolicy:   true,
-			},
-			want: func(tt assert.TestingT, i1 interface{}, i2 ...interface{}) bool {
-				d, ok := i1.(*azureDiscovery)
-				if !assert.True(tt, ok) {
-					return false
-				}
-
-				want := []*voc.Backup{{
-					RetentionPeriod: Duration30Days,
-					Enabled:         true,
-					Storage:         voc.ResourceID("/subscriptions/00000000-0000-0000-0000-000000000000/resourceGroups/res1/providers/Microsoft.DataProtection/backupVaults/backupAccount1/backupInstances/disk1-disk1-22222222-2222-2222-2222-222222222222"),
-					TransportEncryption: &voc.TransportEncryption{
-						Enforced:   true,
-						Enabled:    true,
-						TlsVersion: constants.TLS1_2,
-						Algorithm:  constants.TLS,
-					},
-				},
-				}
-
-				return assert.Equal(t, want, d.backupMap[DataSourceTypeDisc].backup["/subscriptions/00000000-0000-0000-0000-000000000000/resourceGroups/res1/providers/Microsoft.Compute/disks/disk1"])
-			},
-			wantErr: assert.NoError,
-		},
-	}
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			// d := &azureDiscovery{
-			// 	azureDiscovery: tt.fields.azureDiscovery,
-			// }
-			d := tt.fields.azureDiscovery
-
-			// Set clients if needed
-			if tt.fields.clientBackupVault {
-				// initialize backup vaults client
-				_ = d.initBackupVaultsClient()
-			}
-			if tt.fields.clientBackupInstance {
-				// initialize backup instances client
-				_ = d.initBackupInstancesClient()
-			}
-
-			if tt.fields.clientBackupPolicy {
-				// initialize backup policies client
-				_ = d.initBackupPoliciesClient()
-			}
-
-			err := d.discoverBackupVaults()
-
-			tt.wantErr(t, err)
-
-			if tt.want != nil {
-				tt.want(t, d)
-			}
-		})
-	}
-}
-
 func Test_retentionDuration(t *testing.T) {
 	type args struct {
 		retention string
@@ -876,7 +641,6 @@ func Test_retentionDuration(t *testing.T) {
 func Test_azureDiscovery_discoverDefender(t *testing.T) {
 	type fields struct {
 		azureDiscovery *azureDiscovery
-		clientDefender bool
 	}
 	tests := []struct {
 		name    string
@@ -885,21 +649,9 @@ func Test_azureDiscovery_discoverDefender(t *testing.T) {
 		wantErr assert.ErrorAssertionFunc
 	}{
 		{
-			name: "defenderClient not set",
-			fields: fields{
-				azureDiscovery: NewMockAzureDiscovery(newMockStorageSender()),
-				clientDefender: false,
-			},
-			want: nil,
-			wantErr: func(t assert.TestingT, err error, i ...interface{}) bool {
-				return assert.ErrorContains(t, err, "defenderClient not set")
-			},
-		},
-		{
 			name: "Happy path",
 			fields: fields{
 				azureDiscovery: NewMockAzureDiscovery(newMockStorageSender()),
-				clientDefender: true,
 			},
 			want: func(tt assert.TestingT, i1 interface{}, i2 ...interface{}) bool {
 				got, ok := i1.(map[string]*defenderProperties)
@@ -919,17 +671,7 @@ func Test_azureDiscovery_discoverDefender(t *testing.T) {
 	}
 	for _, tt := range tests {
 		t.Run(tt.name, func(t *testing.T) {
-			// d := &azureDiscovery{
-			// 	azureDiscovery: tt.fields.azureDiscovery,
-			// }
-
 			d := tt.fields.azureDiscovery
-
-			// Set client if needed
-			if tt.fields.clientDefender {
-				// initialize backup vaults client
-				_ = d.initDefenderClient()
-			}
 
 			got, err := d.discoverDefender()
 
@@ -937,279 +679,6 @@ func Test_azureDiscovery_discoverDefender(t *testing.T) {
 
 			if tt.want != nil {
 				tt.want(t, got)
-			}
-		})
-	}
-}
-
-// TODO(anatheka): Update test
-func Test_azureDiscovery_discoverBackupInstances(t *testing.T) {
-	type fields struct {
-		azureDiscovery       *azureDiscovery
-		clientBackupInstance bool
-	}
-	type args struct {
-		resourceGroup string
-		vaultName     string
-	}
-	tests := []struct {
-		name    string
-		fields  fields
-		args    args
-		want    []*armdataprotection.BackupInstanceResource
-		wantErr assert.ErrorAssertionFunc
-	}{
-		{
-			name: "Input empty",
-			args: args{},
-			want: nil,
-			wantErr: func(t assert.TestingT, err error, i ...interface{}) bool {
-				return assert.ErrorContains(t, err, "missing resource group and/or vault name")
-			},
-		},
-		{
-			name: "defenderClient not set",
-			fields: fields{
-				azureDiscovery:       NewMockAzureDiscovery(newMockNetworkSender()),
-				clientBackupInstance: true,
-			},
-			args: args{
-				resourceGroup: "res1",
-				vaultName:     "backupAccount1",
-			},
-			want: nil,
-			wantErr: func(t assert.TestingT, err error, i ...interface{}) bool {
-				return assert.ErrorContains(t, err, "error getting next page: GET")
-			},
-		},
-		{
-			name: "Happy path",
-			fields: fields{
-				azureDiscovery:       NewMockAzureDiscovery(newMockStorageSender()),
-				clientBackupInstance: true,
-			},
-			args: args{
-				resourceGroup: "res1",
-				vaultName:     "backupAccount1",
-			},
-			wantErr: assert.NoError,
-			want: []*armdataprotection.BackupInstanceResource{
-				{
-					ID:   util.Ref("/subscriptions/00000000-0000-0000-0000-000000000000/resourceGroups/res1/providers/Microsoft.DataProtection/backupVaults/backupAccount1/backupInstances/account1-account1-22222222-2222-2222-2222-222222222222"),
-					Name: util.Ref("account1-account1-22222222-2222-2222-2222-222222222222"),
-					Properties: &armdataprotection.BackupInstance{
-						DataSourceInfo: &armdataprotection.Datasource{
-							ResourceID:     util.Ref("/subscriptions/00000000-0000-0000-0000-000000000000/resourceGroups/res1/providers/Microsoft.Storage/storageAccounts/account1"),
-							DatasourceType: util.Ref("Microsoft.Storage/storageAccounts/blobServices"),
-						},
-						PolicyInfo: &armdataprotection.PolicyInfo{
-							PolicyID: util.Ref("/subscriptions/00000000-0000-0000-0000-000000000000/resourceGroups/res1/providers/Microsoft.DataProtection/backupVaults/backupAccount1/backupPolicies/backupPolicyContainer"),
-						},
-					},
-				},
-			},
-		},
-	}
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			// d := &azureDiscovery{
-			// 	azureDiscovery: tt.fields.azureDiscovery,
-			// }
-
-			d := tt.fields.azureDiscovery
-
-			if tt.fields.clientBackupInstance {
-				// initialize backup instances client
-				_ = d.initBackupInstancesClient()
-			}
-			got, err := d.discoverBackupInstances(tt.args.resourceGroup, tt.args.vaultName)
-
-			tt.wantErr(t, err)
-
-			if !reflect.DeepEqual(got, tt.want) {
-				t.Errorf("azureDiscovery.discoverBackupInstances() = %v, want %v", got, tt.want)
-			}
-		})
-	}
-}
-
-// TODO(anatheka): Update test
-func Test_azureDiscovery_handleInstances(t *testing.T) {
-	type fields struct {
-		isAuthorized        bool
-		sub                 *armsubscription.Subscription
-		cred                azcore.TokenCredential
-		rg                  *string
-		clientOptions       arm.ClientOptions
-		discovererComponent string
-		clients             clients
-		csID                string
-		backupMap           map[string]*backup
-	}
-	type args struct {
-		vault    *armdataprotection.BackupVaultResource
-		instance *armdataprotection.BackupInstanceResource
-	}
-	tests := []struct {
-		name         string
-		fields       fields
-		args         args
-		wantResource voc.IsCloudResource
-		wantErr      assert.ErrorAssertionFunc
-	}{
-		{
-			name: "Empty input",
-			wantErr: func(t assert.TestingT, err error, i ...interface{}) bool {
-				return assert.ErrorContains(t, err, ErrVaultInstanceIsEmpty.Error())
-			},
-		},
-		{
-			name: "Happy path: ObjectStorage",
-			fields: fields{
-				csID: testdata.MockCloudServiceID1,
-			},
-			args: args{
-				vault: &armdataprotection.BackupVaultResource{
-					ID:       util.Ref("/subscriptions/00000000-0000-0000-0000-000000000000/resourceGroups/res1/providers/Microsoft.DataProtection/backupVaults/backupAccount1"),
-					Name:     util.Ref("backupAccount1"),
-					Location: util.Ref("westeurope"),
-				},
-				instance: &armdataprotection.BackupInstanceResource{
-					ID:   util.Ref("/subscriptions/00000000-0000-0000-0000-000000000000/resourceGroups/res1/providers/Microsoft.DataProtection/backupVaults/backupAccount1/backupInstances/account1-account1-22222222-2222-2222-2222-222222222222"),
-					Name: util.Ref("account1-account1-22222222-2222-2222-2222-222222222222"),
-					Properties: &armdataprotection.BackupInstance{
-						DataSourceInfo: &armdataprotection.Datasource{
-							DatasourceType: util.Ref("Microsoft.Storage/storageAccounts/blobServices"),
-						},
-					},
-				},
-			},
-			wantResource: &voc.ObjectStorage{
-				Storage: &voc.Storage{
-					Resource: &voc.Resource{
-						ID:        "/subscriptions/00000000-0000-0000-0000-000000000000/resourceGroups/res1/providers/Microsoft.DataProtection/backupVaults/backupAccount1/backupInstances/account1-account1-22222222-2222-2222-2222-222222222222",
-						ServiceID: testdata.MockCloudServiceID1,
-						Name:      "account1-account1-22222222-2222-2222-2222-222222222222",
-						GeoLocation: voc.GeoLocation{
-							Region: "westeurope",
-						},
-						CreationTime: 0,
-						Type:         voc.ObjectStorageType,
-						Labels:       nil,
-						Parent:       voc.ResourceID("/subscriptions/00000000-0000-0000-0000-000000000000/resourceGroups/res1"),
-						Raw:          "{\"*armdataprotection.BackupInstanceResource\":[{\"properties\":{\"dataSourceInfo\":{\"datasourceType\":\"Microsoft.Storage/storageAccounts/blobServices\"}},\"id\":\"/subscriptions/00000000-0000-0000-0000-000000000000/resourceGroups/res1/providers/Microsoft.DataProtection/backupVaults/backupAccount1/backupInstances/account1-account1-22222222-2222-2222-2222-222222222222\",\"name\":\"account1-account1-22222222-2222-2222-2222-222222222222\"}],\"*armdataprotection.BackupVaultResource\":[{\"id\":\"/subscriptions/00000000-0000-0000-0000-000000000000/resourceGroups/res1/providers/Microsoft.DataProtection/backupVaults/backupAccount1\",\"location\":\"westeurope\",\"name\":\"backupAccount1\"}]}",
-					},
-				},
-			},
-			wantErr: assert.NoError,
-		},
-		{
-			name: "Happy path: BlockStorage",
-			fields: fields{
-				csID: testdata.MockCloudServiceID1,
-			},
-			args: args{
-				vault: &armdataprotection.BackupVaultResource{
-					ID:       util.Ref("/subscriptions/00000000-0000-0000-0000-000000000000/resourceGroups/res1/providers/Microsoft.DataProtection/backupVaults/backupAccount1"),
-					Name:     util.Ref("backupAccount1"),
-					Location: util.Ref("westeurope"),
-				},
-				instance: &armdataprotection.BackupInstanceResource{
-					ID:   util.Ref("/subscriptions/00000000-0000-0000-0000-000000000000/resourceGroups/res1/providers/Microsoft.DataProtection/backupVaults/backupAccount1/backupInstances/disk1-disk1-22222222-2222-2222-2222-222222222222"),
-					Name: util.Ref("disk1-disk1-22222222-2222-2222-2222-222222222222"),
-					Properties: &armdataprotection.BackupInstance{
-						DataSourceInfo: &armdataprotection.Datasource{
-							DatasourceType: util.Ref("Microsoft.Compute/disks"),
-						},
-					},
-				},
-			},
-			wantResource: &voc.BlockStorage{
-				Storage: &voc.Storage{
-					Resource: &voc.Resource{
-						ID:        "/subscriptions/00000000-0000-0000-0000-000000000000/resourceGroups/res1/providers/Microsoft.DataProtection/backupVaults/backupAccount1/backupInstances/disk1-disk1-22222222-2222-2222-2222-222222222222",
-						ServiceID: testdata.MockCloudServiceID1,
-						Name:      "disk1-disk1-22222222-2222-2222-2222-222222222222",
-						GeoLocation: voc.GeoLocation{
-							Region: "westeurope",
-						},
-						CreationTime: 0,
-						Type:         voc.BlockStorageType,
-						Labels:       nil,
-						Parent:       voc.ResourceID("/subscriptions/00000000-0000-0000-0000-000000000000/resourceGroups/res1"),
-						Raw:          "{\"*armdataprotection.BackupInstanceResource\":[{\"properties\":{\"dataSourceInfo\":{\"datasourceType\":\"Microsoft.Compute/disks\"}},\"id\":\"/subscriptions/00000000-0000-0000-0000-000000000000/resourceGroups/res1/providers/Microsoft.DataProtection/backupVaults/backupAccount1/backupInstances/disk1-disk1-22222222-2222-2222-2222-222222222222\",\"name\":\"disk1-disk1-22222222-2222-2222-2222-222222222222\"}],\"*armdataprotection.BackupVaultResource\":[{\"id\":\"/subscriptions/00000000-0000-0000-0000-000000000000/resourceGroups/res1/providers/Microsoft.DataProtection/backupVaults/backupAccount1\",\"location\":\"westeurope\",\"name\":\"backupAccount1\"}]}",
-					},
-				},
-			},
-			wantErr: assert.NoError,
-		},
-	}
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			d := &azureDiscovery{
-				isAuthorized:        tt.fields.isAuthorized,
-				sub:                 tt.fields.sub,
-				cred:                tt.fields.cred,
-				rg:                  tt.fields.rg,
-				clientOptions:       tt.fields.clientOptions,
-				discovererComponent: tt.fields.discovererComponent,
-				clients:             tt.fields.clients,
-				csID:                tt.fields.csID,
-				backupMap:           tt.fields.backupMap,
-			}
-			gotResource, err := d.handleInstances(tt.args.vault, tt.args.instance)
-			tt.wantErr(t, err)
-
-			assert.Equal(t, tt.wantResource, gotResource)
-		})
-	}
-}
-
-// TODO(anatheka): Update test
-func Test_backupsEmptyCheck(t *testing.T) {
-	type args struct {
-		backups []*voc.Backup
-	}
-	tests := []struct {
-		name string
-		args args
-		want []*voc.Backup
-	}{
-		{
-			name: "Happy path",
-			args: args{
-				backups: []*voc.Backup{
-					{
-						Enabled:         true,
-						Interval:        90,
-						RetentionPeriod: 100,
-					},
-				},
-			},
-			want: []*voc.Backup{
-				{
-					Enabled:         true,
-					Interval:        90,
-					RetentionPeriod: 100,
-				},
-			},
-		},
-		{
-			name: "Happy path: empty input",
-			args: args{},
-			want: []*voc.Backup{
-				{
-					Enabled:         false,
-					RetentionPeriod: -1,
-					Interval:        -1,
-				},
-			},
-		},
-	}
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			if got := backupsEmptyCheck(tt.args.backups); !reflect.DeepEqual(got, tt.want) {
-				t.Errorf("backupsEmptyCheck() = %v, want %v", got, tt.want)
 			}
 		})
 	}
